4 osio valmis

+++ osa4/blogilista/server/controllers/blogs.js
@@ -1,26 +1,69 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const config = require('../utils/config')
 require('express-async-errors')
 
+
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
+
   response.json(blogs.map(blog => blog.toJSON()))
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
 
-  if (!blog.likes) {
-    blog.likes = 0
+  const body = request.body
+  const token = request.token
+
+  const decodedToken = jwt.verify(token, config.SECRET)
+  if (!request || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid ' })
   }
 
-  const result = await blog.save()
-  response.status(201).json(result)
+  const user = await User.findById(decodedToken.id)
+
+  if (!body.likes) {
+    body.likes = 0
+  }
+
+  const blog = new Blog(
+    {
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes,
+      user: user._id
+    }
+  )
+
+  const savedBlog = await blog.save()
+
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+  response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const token = request.token
+
+  const decodedToken = jwt.verify(token, config.SECRET)
+  if (!request || !decodedToken) {
+    return response.status(401).json( { error: 'token missing or invalid ' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if ( blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(401).json( { error: 'permission denied ' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
